pyproject.toml
[project]
name = "rlm"
version = "0.1.0"
description = "Recursive Language Models implementation using Gemini"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai",
    "pymupdf",
    "pypdf>=6.6.0",
    "python-docx",
    "python-dotenv",
    "python-pptx>=1.0.2",
    "restrictedpython>=8.1",
    "rich",
    "termcolor",
    "textual",
]

[project.scripts]
rlm = "rlm.cli.main:main"

[tool.uv]
dev-dependencies = [
    "pytest",
]

[tool.isort]
profile = "black"

[tool.black]
line-length = 88

